Move functions out of src/headless/core.js into utils
